pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llps-diffusion-predictor"
version = "0.1.0"
description = "Open-source framework for protein LLPS prediction with diffusion models."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "LLPS Diffusion Contributors" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "matplotlib>=3.9",
  "pyyaml>=6.0.1",
  "torch>=2.2",
  "scikit-learn>=1.5",
  "tqdm>=4.66",
  "requests>=2.32",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.6.0",
  "mypy>=1.10",
  "pre-commit>=3.7",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
pretty = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --disable-warnings"

